Set python2 path configuration file conditionally
GR 3.10.0 no longer depends on python2 (nor installs it). Hence, it is
not necessary to set up the path configuration file for python2.
Besides, an attempt to set up the .pth file for python2 would fail due
to python2 being unavailable. So the solution is to execute this step
conditionally on python2's availability.